separated vCardParamMap and vCardProperty into their own files
